requirements.txt
playwright
pytest-playwright
pytest
allure-pytest
python-decouple
pytest-xdist